[chore] package rename Camel case -> snake case
